Extract GradientTile from ContactTile so Student Orgs can share it
